% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaned_weather.R, R/lesion_counts.R,
%   R/summary_weather.R
\docType{data}
\name{cleaned_weather}
\alias{cleaned_weather}
\title{Cleaned Weather Data from Horsham and Curyo}
\format{
A \code{data.frame} with 2202 rows and 6 variables:
\describe{
  \item{site}{Location where trial was located}
  \item{rep}{Replicate number. The sequential number of rain or irrigation
   event for a given site}
  \item{time}{Date in POSIXct, format YYYY-MM-DD HH:MM:SS}
  \item{wind_speed}{Recorded wind speed in metres per second}
  \item{wind_direction}{Recorded wind direction in compass degrees}
  \item{rainfall}{Recorded precipitation in millimetres}
}

A \code{data.frame} with 336 rows and 17 variables:
\describe{
  \item{SpEv}{Spread event. An identifier that includes the site and
    replicate number}
  \item{site}{Location where trial was located}
  \item{rep}{Replicate number. The sequential number of rain or irrigation
   event for a given site}
  \item{distance}{Distance of station locations along transect from infected
   plot source in metres}
  \item{station}{Number assigned to each respective station at a given
   distance on a given transect}
  \item{transect}{Number identifying one of ten transects upon which the
   stations were placed}
  \item{plant_no}{Number identifying a given plant in a pot}
  \item{pot_no}{Number identifying a given pot}
  \item{counts_p1}{Total recorded number of lesions in a given pot}
  \item{counts_p2}{Total recorded number of lesions in a given pot}
  \item{counts_p3}{Total recorded number of lesions in a given pot}
  \item{counts_p4}{Total recorded number of lesions in a given pot}
  \item{counts_p5}{Total recorded number of lesions in a given pot}
  \item{degrees}{Compass degrees for a given transect}
  \item{ptype}{Type of precipitation for an event, one of three, 
  \code{rainfall}, \code{irrigation} or \code{mixed}}
  \item{mean_pot_count}{Mean value of lesion counts per pot for that row of
   observations}
}

A \code{data.frame} with 2202 rows and 6 variables:
\describe{
  \item{site}{Location where trial was located}
  \item{rep}{Replicate number. The sequential number of rain or irrigation
   event for a given site}
  \item{mws}{Mean wind speed for the event}
  \item{ws_sd}{Standard deviation of the wind speed for the event}
  \item{mwd}{Mean recorded wind direction in compass degrees as calcuated by
   \code{circular.averaging()}}
  \item{sum_rain}{Sum total of recorded precipitation for the event in
   millimetres}
}
}
\source{
Dr Joshua Fanning, AgricultureVictoria, Horsham, Vic, AUS

GRDC Project USQ1903-003-RTX

Dr Joshua Fanning, AgricultureVictoria, Horsham, Vic, AUS
}
\usage{
cleaned_weather

cleaned_weather

cleaned_weather
}
\description{
A dataset of ten-minute observation weather data from on-site weather
 stations at Horsham and Curyo.

A dataset of summarised chickpea ascochyta blight lesions from experimental
plots located at Horsham and Curyo.

A summarised dataset of weather data from on-site weather stations at Horsham
 and Curyo.
}
\keyword{datasets}
