% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChickpeaAscoDispersal-package.R
\docType{package}
\name{ChickpeaAscoDispersal-package}
\alias{ChickpeaAscoDispersal}
\alias{ChickpeaAscoDispersal-package}
\title{ChickpeaAscoDispersal: Research Compendium for the Paper 'The role of conidia in the dispersal of Ascochyta rabiei'}
\description{
Documents the workflow and processes involved in the analysis of
   data for the paper 'The role of conidia in the dispersal of Ascochyta
   rabiei'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adamhsparks/ChickpeaAscoDispersal}
  \item Report bugs at \url{https://github.com/adamhsparks/ChickpeaAscoDispersal/issues}
}

}
\author{
\strong{Maintainer}: Adam H. Sparks \email{adam.sparks@usq.edu.au} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})

Authors:
\itemize{
  \item Ihsanul Khaliq \email{ihsan.khaliq@usq.edu.au} (\href{https://orcid.org/0000-0003-4171-0917}{ORCID})
  \item Joshua Fanning \email{joshua.fanning@agriculture.vic.gov.au} (\href{https://orcid.org/0000-0002-7730-8498}{ORCID})
  \item Paul Melloy \email{paul.melloy@usq.edu.au} (\href{https://orcid.org/0000-0003-4253-7167}{ORCID})
  \item Jean Galloway \email{jean.galloway@dpird.wa.gov.au}
  \item Kevin Moore \email{kevin.moore@dpi.nsw.gov.au} (\href{https://orcid.org/0000-0001-9643-6885}{ORCID})
  \item Daniel Burrell \email{daniel.burrell@usq.edu.au} (\href{https://orcid.org/0000-0002-9173-651X}{ORCID})
}

Other contributors:
\itemize{
  \item GRDC Project USQ1903-003RTX [funder]
  \item Agriculture Victoria and GRDC co-investment project DAV00150 [funder, data contributor]
  \item GRDC Project DJP1097-001RTX [funder]
  \item The Birchip Cropping Group [data contributor]
}

}
\keyword{internal}
