% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_gam.R
\name{p_gam}
\alias{p_gam}
\title{Generic function to print GAMs using mgcViz}
\usage{
p_gam(x)
}
\arguments{
\item{x}{a GAM object created by `[mgcv]gam()`}
}
\value{
a `ggplot2` graph object of the GAM
}
\description{
Generic function to print GAMs using mgcViz
}
\examples{
# load necessary libraries
library("tidyverse")
library("mgcv")
library("here")
library("mgcViz")
library("ChickpeaAscoDispersal")

# load weather and lesion data and make a simple GAM
lesion <- read_csv(system.file("extdata", "summarised_lesion_counts.csv",
                               package = "ChickpeaAscoDispersal",
                               mustWork = TRUE))
weather <- read_csv(system.file("extdata", "weather_summary.csv",
                                package = "ChickpeaAscoDispersal",
                                mustWork = TRUE))
dat <- left_join(lesion, weather, by = c("site", "rep"))

mod1 <- gam(
  mean_pot_count ~ s(distance, k = 5),
  data = dat
)

print(p_gam(x = getViz(mod1)) +
        ggtitle("s(Distance)"),
      pages = 1)
      
}
