% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_gam.R
\name{p_gam}
\alias{p_gam}
\title{Generic function to print GAMs using mgcViz}
\usage{
p_gam(x)
}
\arguments{
\item{x}{a GAM object created by \code{\link[mgcv]{gam}}}
}
\value{
a `ggplot2` graph object of the GAM
}
\description{
Generic function to print GAMs using mgcViz
}
\examples{
# load necessary libraries
library("tidyverse")
library("mgcv")
library("mgcViz")
library("ChickpeaAscoDispersal")

# these data are automatically loaded
dat <- left_join(lesion_counts, summary_weather, by = c("site", "rep"))

mod1 <- gam(
  m_lesions ~ s(distance, k = 5),
  data = dat
)

print(p_gam(x = getViz(mod1)) +
        ggtitle("s(Distance)"),
      pages = 1)
      
}
