% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_gam.R
\name{p_gam}
\alias{p_gam}
\title{Generic function to print GAMs using mgcViz}
\usage{
p_gam(x)
}
\arguments{
\item{x}{a GAM object created by `[mgcv]gam()`}
}
\value{
a `ggplot2` graph object of the GAM
}
\description{
Generic function to print GAMs using mgcViz
}
\examples{
\donttest{
library(mgcv)
set.seed(2) ## simulate some data...
dat <- gamSim(1,n=400,dist="normal",scale=2)
b <- gam(y~s(x0)+s(x1)+s(x2)+s(x3),data=dat)
p_gam(b)
}
}
